Fix a freeze when pasting lots of text
