+++ src/cascadia/TerminalApp/TerminalPage.cpp
@@ -8,6 +8,7 @@
 #include <TerminalCore/ControlKeyStates.hpp>
 #include <TerminalThemeHelpers.h>
 #include <til/hash.h>
+#include <til/unicode.h>
 #include <Utils.h>
 
 #include "../../types/inc/ColorFix.hpp"
@@ -2996,7 +2997,19 @@ namespace winrt::TerminalApp::implementation
             {
                 // We have to initialize the dialog here to be able to change the text of the text block within it
                 std::ignore = FindName(L"MultiLinePasteDialog");
-                ClipboardText().Text(text);
+
+                // WinUI absolutely cannot deal with large amounts of text (at least O(n), possibly O(n^2),
+                // so we limit the string length here and add an ellipsis if necessary.
+                auto clipboardText = text;
+                if (clipboardText.size() > 1024)
+                {
+                    const std::wstring_view view{ text };
+                    // Make sure we don't cut in the middle of a surrogate pair
+                    const auto len = til::utf16_iterate_prev(view, 512);
+                    clipboardText = til::hstring_format(FMT_COMPILE(L"{}\n…"), view.substr(0, len));
+                }
+
+                ClipboardText().Text(std::move(clipboardText));
 
                 // The vertical offset on the scrollbar does not reset automatically, so reset it manually
                 ClipboardContentScrollViewer().ScrollToVerticalOffset(0);
@@ -3012,7 +3025,7 @@ namespace winrt::TerminalApp::implementation
                 }
 
                 // Clear the clipboard text so it doesn't lie around in memory
-                ClipboardText().Text(L"");
+                ClipboardText().Text({});
 
                 if (warningResult != ContentDialogResult::Primary)
                 {
