+++ src/til/ut_til/til.unit.tests.vcxproj
@@ -45,10 +45,13 @@
     <ClInclude Include="..\..\inc\til\bytes.h" />
     <ClInclude Include="..\..\inc\til\coalesce.h" />
     <ClInclude Include="..\..\inc\til\color.h" />
+    <ClInclude Include="..\..\inc\til\colorbrewer.h" />
     <ClInclude Include="..\..\inc\til\enumset.h" />
     <ClInclude Include="..\..\inc\til\env.h" />
+    <ClInclude Include="..\..\inc\til\flat_set.h" />
     <ClInclude Include="..\..\inc\til\generational.h" />
     <ClInclude Include="..\..\inc\til\hash.h" />
+    <ClInclude Include="..\..\inc\til\io.h" />
     <ClInclude Include="..\..\inc\til\latch.h" />
     <ClInclude Include="..\..\inc\til\math.h" />
     <ClInclude Include="..\..\inc\til\mutex.h" />
@@ -57,6 +60,7 @@
     <ClInclude Include="..\..\inc\til\point.h" />
     <ClInclude Include="..\..\inc\til\rand.h" />
     <ClInclude Include="..\..\inc\til\rect.h" />
+    <ClInclude Include="..\..\inc\til\regex.h" />
     <ClInclude Include="..\..\inc\til\replace.h" />
     <ClInclude Include="..\..\inc\til\rle.h" />
     <ClInclude Include="..\..\inc\til\size.h" />
@@ -69,6 +73,7 @@
     <ClInclude Include="..\..\inc\til\type_traits.h" />
     <ClInclude Include="..\..\inc\til\u8u16convert.h" />
     <ClInclude Include="..\..\inc\til\unicode.h" />
+    <ClInclude Include="..\..\inc\til\winrt.h" />
     <ClInclude Include="..\precomp.h" />
   </ItemGroup>
   <ItemDefinitionGroup>
